+++ packages/sodium/CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.2] - 2021-05-12
+## [0.1.3] - 2021-05-21
 ### Added
-- New libsodium API: crypto_box
+- New libsodium APIs:
+  - crypto_box
+  - crypto_sign
+### Changed
+- Set minimum required dart version to 2.13.0
 
 ## [0.1.1] - 2021-05-11
 ### Added